requirements.txt
whisper
torch
librosa
numpy
soundfile
psutil
transformers



edge-tts
sentencepiece
google-genai
python-dotenv



sentence_transformers
langchain
langchain_community
tiktoken 
pdfminer.six


streamlit
